REFACTOR & FIX - Background Container & Painter Resource Saver
